fixed faq fitting for how-it-works page and fixed sizing of a button in index
